Home HTML CSS Clock update 5/8/2021
Using a guide, I  went ahead and created a real time clock that we will use to calculate meeting times

NOTE: I followed the guide, but made a mistake somewhere which didn't displayed the clock at all, I found out, however, that the video had a link to the source code, which I used to remove any other bugs I might've overlooked. The link to both the video and the source code is in the comments if you want to reference it, there are other codes we can see in the website that might be good for the project.