cyberComp Table 5/9/2021
So I editing the images on the Cyber Comp tab to use up more spaces, but the problem now that I found out is that the div class on home css, "clock" actually makes whatever affected to stay in place, and by that I mean that if you were to shrink the website, the images stay static without properly aligning. I'm gonna have to look for a proper solution for the div class. This is just a note in case you're gonna change the code in some way.