Home HTML 5/14/2021
So I realized that if I were to try to make the Javascript countdown formula, I would have to basically edit the scripts that were not make for that.

I'll still try to figure out how to make the formula, but till then, I make a simplified solution, which just straight up tells the user when the meetings will be. This will be the solution until I figure something else out.