Home css html, and cyberComp HTML
For the minor things, I swapped the images in cyber comp with the NCL, as suggested by the profesor, as well as adding the grid-container class from the contact css file.

Major changes in Home html include the grids in the introduction, cyber competitions, and the clock portions of the page, as for the clock part of the page, made it so that not only the spaces would be used more within the text, but the link to the zoom meeting will disappear when it is not meeting day, nor when the meeting has passed, will see if I can add the speech bubble in the introduction later.